Quiz-1-Stalgo-main/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0
